The transmission of files now works properly
changed the fileInfoBuffer to be of size 512 and only the necessary bytes
are written when sending file information. This lets us retrieve the
fitting size one the other end so we dont need to worry about leftover
bytes from the previous transmission (since we only take the number of
bytes read out from the buffer).
Files are seperated by knowing the size of the file (which is now included
in the fileInfo part, as bytes 1-9).
Commented out the methods for only-files transmission in the
NetworkFileReceiver/Sender since those do not have the changes to run yet.
Will probably bring those up to date somewhere along the line.
Transformation of int/long to byteArrays and back is no handled by the
Utils class via static methods.
All functions for the Server are now enabled and are not run seperately
like before.
Several important parameters are now handed down from the main class so
that changing those can be done from one place. Will probably fetch those
from a ini file or something in the future.

+++ src/app_server/NetworkFileReceiver.java
@@ -7,9 +7,12 @@ import java.io.InputStream;
 
 public class NetworkFileReceiver {
 	
+	private final int fileInfoBufferSize = 512;
+	private final int sizeStartOffset = 1;
+	private final int pathStartIndex = 9;
 	private InputStream inStream;
 	private byte[] readBuffer;
-	private byte[] fileInfoBuffer = new byte[128];
+	private byte[] fileInfoBuffer = new byte[fileInfoBufferSize];
 	private int readBytes;
 	
 	public NetworkFileReceiver(InputStream in, int bufferSize) throws IOException{
@@ -17,25 +20,70 @@ public class NetworkFileReceiver {
 		readBuffer = new byte[bufferSize];
 	}
 	
-	public void receiveAndWriteFiles() throws IOException{
-		inStream.read(readBuffer);
-		int numberOfFiles = 0;
-		for (int byteToIntOffset = 0; byteToIntOffset < 4; byteToIntOffset++) {
-	        int shift = (4 - 1 - byteToIntOffset) * 8;
-	        numberOfFiles += (readBuffer[byteToIntOffset] & 0x000000FF) << shift;
-	    }
+//	public void receiveAndWriteFiles(String pathPrefix) throws IOException{
+//		inStream.read(readBuffer);
+//		int numberOfFiles = 0;
+//		for (int byteToIntOffset = 0; byteToIntOffset < 4; byteToIntOffset++) {
+//	        int shift = (4 - 1 - byteToIntOffset) * 8;
+//	        numberOfFiles += (readBuffer[byteToIntOffset] & 0x000000FF) << shift;
+//	    }
+//		File currentFile;
+//		for(int outerCounter = 0; outerCounter < numberOfFiles; outerCounter++){
+//			inStream.read(fileInfoBuffer);
+//			String currentFileName = new String(fileInfoBuffer, "UTF-8");
+//			currentFileName = pathPrefix + currentFileName;
+//			//TODO: testing shit
+////			System.out.println(currentFileName);
+////			currentFileName = "F://Tmp/testFile.pdf";
+//			currentFile = new File(currentFileName);
+//			File parentFile = currentFile.getParentFile();
+//			if(!parentFile.exists() && !parentFile.mkdirs()){
+//			    throw new IllegalStateException("Couldn't create dir: " + parentFile);
+//			}
+//			try(FileOutputStream fileOutStream = new FileOutputStream(currentFile)){
+//				while((readBytes = inStream.read(readBuffer)) > 0){
+//					fileOutStream.write(readBuffer, 0, readBytes);
+//				}
+//			}
+//		}
+//	}
+	
+	public void receiveAndWriteFilesFromDeltas(String pathPrefix) throws IOException{
+		inStream.read(readBuffer, 0, 4);
+		int testIndex = 0;
+		if(pathPrefix.charAt(pathPrefix.length() - 1) != '/'){
+			pathPrefix += "/";
+		}
+		int numberOfFiles = Utils.byteToInt(readBuffer, 0);
+		System.out.println(numberOfFiles);
 		File currentFile;
 		for(int outerCounter = 0; outerCounter < numberOfFiles; outerCounter++){
-			inStream.read(fileInfoBuffer);
-			String currentFileName = new String(fileInfoBuffer, "UTF-8");
-			//TODO: testing shit
-			System.out.println(currentFileName);
-			currentFileName = "F://Tmp/testFile.pdf";
+			readBytes = inStream.read(fileInfoBuffer);
+			long currentFileSize = Utils.byteToLong(fileInfoBuffer, sizeStartOffset);
+			System.out.println(currentFileSize);
+			String currentFileName = new String(fileInfoBuffer, "UTF-8").substring(pathStartIndex, readBytes);
+			currentFileName = pathPrefix + currentFileName;
+			currentFileName = currentFileName.replace("\\", "/");
+			if(testIndex == 0){
+				currentFileName = "F:\\Test\\textures\\fx\\flare_refl_rainbow.tex";
+				testIndex++;
+			}
 			currentFile = new File(currentFileName);
-			try(FileOutputStream fileOutStream = new FileOutputStream(currentFile)){
-				while((readBytes = inStream.read(readBuffer)) > 0){
-					fileOutStream.write(readBuffer, 0, readBytes);
+			File parentFile = currentFile.getParentFile();
+			if(!parentFile.exists() && !parentFile.mkdirs()){
+			    throw new IllegalStateException("Couldn't create dir: " + parentFile);
+			}
+			System.out.println(currentFile.getAbsolutePath());
+			if(fileInfoBuffer[0] != Delta.FILE_REMOVED){
+				try(FileOutputStream fileOutStream = new FileOutputStream(currentFile)){
+					readBytes = 0;
+					for(int counter = 0; counter < currentFileSize; counter = counter + readBytes){
+						readBytes = inStream.read(readBuffer);
+						fileOutStream.write(readBuffer, 0, readBytes);
+					}
 				}
+			} else{
+				currentFile.delete();
 			}
 		}
 	}
